i655 UBG exports python code that can generate the CEA files
It is not easy to install the required libraries in GHpython at
Grasshopper. So I think it is better we run everything outside to avoid
this library issue. Thus, now, the UBG at grasshopper will generate .py
code that can be run to generate the CEA files. Enclosed in the
"temporary" folder are the .py code generated by the UBG.