1263-more changes to docs bat/config and build errors fixed
I created another line in the batch to delete all rst files under the title "cea" to ensure no duplicates are created when modules are merged.

Attempted to fix all menial build errors however some work still needs to be done.

Importing module errors is still a big issue and cannot be addressed in the config.