changed structure of system_costs
I have changed the capital expenditure calculation to be based on the thermal capacity as cost data is mostly given in $/kW thermal capacity.
I further changed the structure of the cost calculation to a more intuitive approach, splitting the script in calculations based on the thermal capacity and the energy consumption of feedstocks.