Add temporary directory as instance variable in inputlocator
This ensures that a new unique temporary directory is created every time it is called within a script. Storing it as an instance variable prevents the temporary directory to be cleared when the tempfile object is garbage collected after it is returned from the method. Since a new inputlocator object is created at the start of a script and garbage collected at the end, this temporary directory should only live for the lifetime of the script.