Temporarily comment out bug-fixes to check refactoring
This part of the code fixes a bug where the storage might have been charged or discharged in the Trigen-activation sequence, despite the Trigen not being activated due to the minimum operation point not being reached.

As soon it is confirmed, that the refactoring doesn't introduce any changes to the code this bug-fix will be reactivated.