VCC part load factor functions removed
After detecting an error with the PLF-based model of the VCCs, the model was reverted back to the VCC model to the simpler g-value model.
The calculation of the PLF and the capacity of VCCs corresponding to that model, introduced by @emanuel030, (although correct according to ASHRAE 90.1 & https://comnet.org/index.php/382-chillers) are unfortunately no longer in use.
These functions, along with a few other unused functions in the script were therefore deleted.