Fix CRS change
The previous change to these scripts saved the zone and surroundings files in the geographic coordinate systems. This change caused issues in some surface calculations in later scripts.
This change now saves them to the projected coordinated system (with the projection point set to the centroid of the zone), which should rectify the errors in the surface calculations.