Supply System Components Update
This push introduced new classes for each of the components/technologies that can be installed in a supply system.
For now the efficiencies of each of the components are only modelled using a constant efficiency model.
The existing temperature dependent efficiency models will need to be woven into the .operate-method of each component at a later point in time.