correct errors in capacity indicator vectors
Previously to this fix, the mutation and recombination of capacity indicator vectors could effectively lead to a understocked supply systems (i.e. supply systems that could not meet the necessary demand, and would simply neglect part of the demand). This fix makes sure the supply system always offers the necessary capacity and meets demand.