Handle invalid base population surviving
In some instances, the initial population contains individuals. They are kept in the population to not distort the genetic algorithms' recombination operation. If the specific supplySystemStructure makes it hard to find valid individuals, some initial invalid individuals might stay in the population until the end. These changes ensure that none of these individuals are kept in the final selection.