Update databases to match up-to-date formats
The data files in the SLE database were based on an older version of CEA's Singapore-database. This mismatch in database formats could lead to crashes when working with some cases. Updating the database formats to the latest version should fix these issues.

This update will have to be renewed in a few weeks as another big change in the CEA databases is scheduled to be pushed to master soon.