Update supply system operation
Previously every time a category of components was operated the water-filling-principal method would check if the energy demand could be drawn from available energy potentials instead of supplying that energy using the components. This lead to false operating calculations since systems would try to draw energy that actually needed to be rejected instead. This newly introduced change should fix that.