Fetching Cost and GHG emissions from feedstock
This change now makes it so that average cost and green house gas emissions of specific energy carriers are fetched from the individual feedstock tabs instead of the overview and cleans up the overview tab accordingly.
IMPORTANT: This does not make cost and greenhouse gas emissions time-bound and does not introduce a differentiation between buying and selling price!