New model for sewage water calculation
Improved sewage model: the external sewage is calculated using OSM data and the related temperature takes into account losses with the ground and the air. Also corrected mistakes related to the mixing of freshwater and hot water in the zone. The calculated temperature creates a new carrier in the energy_carrier document