Modified script, prioritized selection of data centers in the zone
Compared to the last commit, the script will check what is the building type of the building in the zone, and if it finds a SERVERROOM type, it will select it over industries searched using OSM data. The type SERVERROOM should be manually included on CEA if information on the building type is known.
On the contrary, if no data centers are found, generic industries will be selected and treated as data centers.