New solar component, include PV in the optimisation
The purpose of this branch is include the PV in the optimization, and when successful include the PVT and SC as well. In this commit, changes have been applied to select the new component, however the code still exclude the potential as energy carrier since draws it from the grid.