Take time-series into consideration
This change takes the time series given in the feedstock database into consideration when calculating annualised cost and annual GHG emissions for each of the systems during the optimisation.
WARNING: At the same time, this reintroduces the possibility to sell energy to the system, which might lead to odd systems.