Introduced Solar Collector as separate component, changed solar capacity control to 1
Fixed bug that was causing the optimization to get stuck in a loop without ending the optimization. The error was caused by the limitation on the solar area available. Now the PV and SC compete for the total area before the capacity indicator vector is created for the whole system.