Introduce new passive component: inverter for solar production
Introduced a new component, inverter. In this way solar panels are able to produce energy in DC current, and at a lower voltage. The inverter then converts the flow into a higher voltage and in AC. Moreover, improved the capacity indicator limitation for solar technologies by adding the limitation at each capacity vector creation, not only when components capacities are oversized!