Associated time series to building energy potentials
This sets the indices of the building specific energy potentials to the general time series of the energy system.
Also, the variable name 'pvt_potential' was changed to 'building_potential'. This method was originally implemented to fetch PVT potentials specially, but that has long since been extended to other building-specific energy potentials.