fix absorption chiller selection
The decentralized optimisation script updates the properties of it's chiller model with the required massflowrate of chilled water. This adaptation of chiller properties was only based on the chiller code but not the required sizing, once again leading to undersized absorption chiller configurations, which would lead to extreme temperatures. This change should fix that.