Adapting number of networks to district size
This change relates to the optimisation script's parameter for the maximum number of networks to be installed in a given district. Instead of simply installing a fixed maximum number of networks the input parameter is adapted to increase the number of networks with the number of buildings in a district instead. The new input parameter is called the 'buildings_to_networks_ratio'.