Introduce connectivity vector stabilisation
This new operation ensures that connectivity vectors that encode for the same connectivity state of a district are now forced to take on the same expression. This is done in an attempt to reduce the dimensionality of the problem, hopefully reducing the number of selected optimal solutions and speeding up convergence.