rectifying capacity indicator vector values
Some duplicates could previously not be detected as some CapacityIndicatorVectors (CIV) held values of 0.0 and 1.0 (as floats) and others held 0 and 1 (as ints). With this change capacity indicator values of 0 and 1 will continuously be registered as integer values, allowing for proper comparisons and elimination of duplicates during the genetic optimisation.