Update script to fetch excels instead of kmls
The excel files are much smaller in size compared to kml files (800 KB vs 13 MB in the worse case). Fetching source files are now much faster when running this script.

There are updated TMYx files for some areas so we try to fetch the latest version (the current version if it's latest version does not exist).